requirements.txt
sentence-transformer
scikit-learn
numpy
matplotlib
